requirements.txt
pandas==1.2.4
urllib3==1.23
pytest==7.2.0
requests==2.26.0
numpy==1.21.6
celery==4.4.7
redis==3.5.3
pyyaml==5.4.1
flask==2.1.0
flower==0.9.7
gunicorn==20.1.0
pydantic[dotenv]==1.9.1
mega.py==1.0.8
pytube==12.1.0
youtube_dl==2021.12.17
psutil==5.8.0